Add logging

+++ src/tasks/tasks.service.ts
@@ -18,8 +18,8 @@ export class TasksService {
 
     }
 
-    async getTaskById(id: number): Promise<Task>{
-        const task = await this.taskRepository.findOne(id);
+    async getTaskById(id: number, user: User): Promise<Task>{
+        const task = await this.taskRepository.findOne({where:{id, userId:user.id}});
         if(!task){
             throw new NotFoundException();
         }
@@ -31,23 +31,23 @@ export class TasksService {
         return this.taskRepository.createTask(createTaskDto, user);
     }
 
-    async deleteTaskById(id: number):Promise<void> {
-       const result = await this.taskRepository.delete(id);
+    async deleteTaskById(id: number, user: User):Promise<void> {
+       const result = await this.taskRepository.delete({id,userId:user.id});
         
         if(result.affected === 0){
             throw new NotFoundException();
         }
     }
 
-    async updateTaskStatus(id: number, status: TaskStatus):Promise<Task>{
-        const task = await this.getTaskById(id);
+    async updateTaskStatus(id: number, status: TaskStatus, user: User):Promise<Task>{
+        const task = await this.getTaskById(id, user);
         task.status = status;
         await task.save();
         return task;
     }
 
-    getTasks(filterDto: GetTasksFilterDto):Promise<Task[]>{
-        return this.taskRepository.getTasks(filterDto);
+    getTasks(filterDto: GetTasksFilterDto, user:User):Promise<Task[]>{
+        return this.taskRepository.getTasks(filterDto,user);
     }
